splits the CucumberSerenityRunner into a base class with most of the functionality WITHOUT enforcing the SerenityReporter Plugin in the base class
